# docs/swagger.yaml
basePath: /api
definitions:
  application.Application:
    properties:
      created_at:
        type: string
      id:
        type: string
      job_id:
        type: string
      job_title:
        type: string
      resume_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  application.applyReq:
    properties:
      job_id:
        type: string
      resume_id:
        type: string
    required:
    - job_id
    - resume_id
    type: object
  application.updateStatusReq:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  auth.LoginResult:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/auth.User'
    type: object
  auth.User:
    properties:
      avatar_url:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  auth.loginReq:
    properties:
      code:
        type: string
      phone:
        type: string
      role:
        enum:
        - candidate
        - hr
        - admin
        type: string
    required:
    - code
    - phone
    - role
    type: object
  auth.sendCodeReq:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  interview.Invitation:
    properties:
      attachment_urls:
        items:
          type: string
        type: array
      candidate_name:
        type: string
      candidate_user_id:
        type: string
      company_address:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      hr_name:
        type: string
      hr_user_id:
        type: string
      id:
        type: string
      job_application_id:
        type: string
      job_title:
        type: string
      notes:
        type: string
      scheduled_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  interview.interviewReq:
    properties:
      application_id:
        type: string
      company_address:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      notes:
        type: string
      scheduled_at:
        type: string
    required:
    - application_id
    - company_address
    - scheduled_at
    type: object
  interview.updateInvStatusReq:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/maxfeizi04-cloud/recruitment-platform
  description: 放心招聘平台后端接口文档，支持求职者、招聘者、管理员三端
  title: 放心招聘平台 API
  version: "2.0"
paths:
  /applications:
    post:
      consumes:
      - application/json
      parameters:
      - description: 职位ID和简历ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/application.applyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/application.Application'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 投递职位
      tags:
      - 投递
  /applications/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: 投递ID
        in: path
        name: id
        required: true
        type: string
      - description: 状态值（viewed/accepted/rejected）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/application.updateStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新投递状态
      tags:
      - 投递
  /applications/my:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取我的投递记录
      tags:
      - 投递
  /applications/received:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取收到的投递
      tags:
      - 投递
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 手机号、验证码和角色
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 手机验证码登录
      tags:
      - 认证
  /auth/send-code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 手机号
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.sendCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 发送手机验证码
      tags:
      - 认证
  /chat/user-sig:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取IM UserSig
      tags:
      - 即时通讯
...
  /interviews:
    post:
      consumes:
      - application/json
      parameters:
      - description: 面试信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/interview.interviewReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/interview.Invitation'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 发起面试邀约
      tags:
      - 面试
  /interviews/{id}/navigate:
    get:
      consumes:
      - application/json
      parameters:
      - description: 面试ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取面试导航链接
      tags:
      - 面试
  /interviews/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: 面试ID
        in: path
        name: id
        required: true
        type: string
      - description: 状态值（accepted/declined/reschedule/confirmed）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/interview.updateInvStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新面试状态
      tags:
      - 面试
  /interviews/my:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取我的面试邀约
      tags:
      - 面试
...
  /maps/place-search:
    get:
      consumes:
      - application/json
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 地点搜索
      tags:
      - 地图
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"